tools/bigquery_export/requirements.txt
google-cloud-logging==1.12.1
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
google-api-core==1.14.2
google-api-python-client==1.7.10
google-auth==1.6.3
google-auth-httplib2==0.0.3
google-cloud==0.34.0
google-cloud-bigquery==1.18.0
google-cloud-core==1.0.3
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
httplib2==0.13.1
idna==2.8
protobuf==3.9.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
uritemplate==3.0.0
urllib3==1.25.3